feat: RoomControlletTest 기반코드 추가
- 기존에 있던 `RoomControllerTest` 는 `RoomControllerIntegrationTset` 로 이름 변경
