+++ src/main/java/web/chat/backend/ChatApplication.java
@@ -2,10 +2,8 @@ package web.chat.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EnableJpaRepositories
 public class ChatApplication {
 
 	public static void main(String[] args) {
