+++ src/test/java/web/chat/backend/controller/RoomControllerTest.java
@@ -1,69 +1,70 @@
 package web.chat.backend.controller;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.hamcrest.Matchers.not;
-import static org.hamcrest.Matchers.*;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
-import java.util.Objects;
-
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.jdbc.Sql;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.TestConstructor;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.filter.CharacterEncodingFilter;
 
-import web.chat.backend.exception.NotFoundException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import web.chat.backend.service.MessageService;
+import web.chat.backend.service.RoomService;
 
-/**
- * Created by koseungbin on 2020-08-15
- */
-
-@SpringBootTest
+@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
+@RequiredArgsConstructor
+@WebMvcTest(RoomController.class)
 class RoomControllerTest {
-	private MockMvc mockMvc;
 
-	@Autowired
-	private RoomController roomController;
+	final MockMvc mockMvc;
+	final ObjectMapper objectMapper;
 
-	@BeforeEach
-	void setup() {
-		mockMvc = MockMvcBuilders
-			.standaloneSetup(roomController)
-			.setControllerAdvice(new ExceptionController())
-			.addFilters(new CharacterEncodingFilter("UTF-8", true))
-			.build();
+	@MockBean
+	RoomService roomService;
+
+	@MockBean
+	MessageService messageService;
+
+	@Test
+	void getRooms() throws Exception {
+
+		// given
+
+		// when
+
+		// then
 	}
 
 	@Test
-	@Sql("/test-sql/messages.sql")
-	@DisplayName("ID가 1인 채팅방의 메시지 리스트가 조회되야만 한다")
-	void shouldGetMessageList_inRoomIdWith1() throws Exception {
-		mockMvc.perform(get("/api/rooms/{id}/messages", 1))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.*", not(empty())))
-			.andExpect(jsonPath("$.*", hasSize(1)))
-			.andExpect(jsonPath("$[0].id", is(1)))
-			.andExpect(jsonPath("$[0].contents", is("foo")))
-			.andExpect(jsonPath("$[0].messageType", is("TEXT")))
-			.andDo(print());
+	void createRoom() throws Exception {
+
+		// given
+
+		// when
+
+		// then
 	}
 
+	@DisplayName("Room의 title 길이 2미만 불가능")
 	@Test
-	@DisplayName("존재하지 않은 채팅방을 요청한 경우에는 400 응답코드를 반환해야만 한다")
-	void shouldRespond400StatusCode_ifNotFoundRoom() throws Exception {
-		mockMvc.perform(get("/api/rooms/{id}/messages", 100))
-			.andExpect(status().isBadRequest())
-			.andExpect(result -> assertThat(result.getResolvedException() instanceof NotFoundException).isTrue())
-			.andExpect(result -> assertEquals(Objects.requireNonNull(result.getResolvedException()).getMessage(),
-				"100 does not exist."))
-			.andDo(print());
+	void createRoom_titleLengthLessThan_2() throws Exception {
+
+		// given
+
+		// when
+
+		// then
+	}
+
+	@DisplayName("Room의 title 길이 20초과 불가능")
+	@Test
+	void createRoom_titleLengthExceed_20() throws Exception {
+
+		// given
+
+		// when
+
+		// then
 	}
 }
